# [13.0] Optimize StockBuffer.cron_actions()