@LoisRForgeFlow 2020-11-05
> > However I can count already 5 OCA modules sharing the same implementation of `is_sublocation_of` method, we would need a module for that. Maybe a module including a collection of useful methods related to stock? (I'm thinking about some used in `shopfloor`).
> 
> Yes... I had some hope we could get something like this: [odoo/odoo#53866](https://github.com/odoo/odoo/pull/53866)
> But it won't happen so we should have a base stock module for this.

Yes, I would be great if they finally consider it, it is indeed very useful.